Added the acl and other fixes. Added the api interfaces to work with the module.

+++ Block/Adminhtml/Pdfgenerator/Edit/DeleteButton.php
@@ -17,7 +17,7 @@
  * @license     http://opensource.org/licenses/afl-3.0.php  Academic Free License (AFL 3.0)
  */
 
-namespace  Eadesigndev\Pdfgenerator\Block\Adminhtml\Pdfgenerator\Edit;
+namespace Eadesigndev\Pdfgenerator\Block\Adminhtml\Pdfgenerator\Edit;
 
 use Magento\Framework\View\Element\UiComponent\Control\ButtonProviderInterface;
 
@@ -32,18 +32,20 @@ class DeleteButton extends GenericButton implements ButtonProviderInterface
      */
     public function getButtonData()
     {
-        $data = [];
-        if ($this->getTemplateId()) {
-            $data = [
-                'label' => __('Delete Template'),
-                'class' => 'delete',
-                'on_click' => 'deleteConfirm(\'' . __(
-                    'Are you sure you want to do this?'
-                ) . '\', \'' . $this->getDeleteUrl() . '\')',
-                'sort_order' => 20,
-            ];
+        if ($this->_isAllowedAction(\Eadesigndev\Pdfgenerator\Controller\Adminhtml\Templates::ADMIN_RESOURCE_SAVE)) {
+            $data = [];
+            if ($this->getTemplateId()) {
+                $data = [
+                    'label' => __('Delete Template'),
+                    'class' => 'delete',
+                    'on_click' => 'deleteConfirm(\'' . __(
+                            'Are you sure you want to do this?'
+                        ) . '\', \'' . $this->getDeleteUrl() . '\')',
+                    'sort_order' => 20,
+                ];
+            }
+            return $data;
         }
-        return $data;
     }
 
     /**
